+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Fixed the suggested rate in `check_regular_timestamps` to be in Hz [#467](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/467)
 
+* Added a skip for mac sidecar files (._*): [#470](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/470)
+
 # v0.4.35
 
 ### Fixes